refactor: change description to title
