+++ lib/config/gherlintrc.js
@@ -26,6 +26,6 @@ module.exports = {
         no_then_as_first_step: "warn",
         only_given_step_in_background: "warn",
         newline_before_scenario: "off",
-        lowercase_description: "off",
+        lowercase_title: "off",
     },
 };
