+++ lib/rules/index.js
@@ -10,6 +10,6 @@ module.exports = {
         no_then_as_first_step: require("./no_then_as_first_step"),
         only_given_step_in_background: require("./only_given_step_in_background"),
         newline_before_scenario: require("./newline_before_scenario"),
-        lowercase_description: require("./lowercase_description"),
+        lowercase_title: require("./lowercase_title"),
     },
 };
