+++ tests/lib/rules/lowercase_title.test.js
@@ -3,14 +3,14 @@ const {
     getValidTestData,
     getInvalidTestData,
     getInvalidTestDataWithFix,
-} = require("../../__fixtures__/Rules/lowercase_description/fixture");
-const LowercaseDescription = require("../../../lib/rules/lowercase_description");
+} = require("../../__fixtures__/Rules/lowercase_title/fixture");
+const LowercaseTitle = require("../../../lib/rules/lowercase_title");
 
 const config = {
     type: "off",
 };
 
-describe("lowercase_description", () => {
+describe("lowercase_title", () => {
     describe("invalid ast", () => {
         it.each([
             ["undefined", undefined],
@@ -18,29 +18,29 @@ describe("lowercase_description", () => {
             ["string", ""],
             ["empty object", {}],
         ])("%s: should not show any lint problems", (_, ast) => {
-            const problems = LowercaseDescription.run(ast);
+            const problems = LowercaseTitle.run(ast);
             expect(problems).toEqual([]);
         });
     });
 
-    describe("with a lowercase description", () => {
+    describe("with a lowercase title", () => {
         const testData = getValidTestData();
         it.each(testData)("%s", (_, text, expectedProblems) => {
             const ast = parser.parse(text);
             ast.text = text;
-            const problems = LowercaseDescription.run(ast, config);
+            const problems = LowercaseTitle.run(ast, config);
             expect(problems).toEqual(expectedProblems);
             expect(problems.length).toEqual(0);
         });
     });
 
-    describe("with a uppercase description", () => {
+    describe("with a uppercase title", () => {
         const testData = getInvalidTestData();
 
         it.each(testData)("%s", (_, text, expectedProblems) => {
             const ast = parser.parse(text);
             ast.text = text;
-            const problems = LowercaseDescription.run(ast, config);
+            const problems = LowercaseTitle.run(ast, config);
             expect(problems.length).toEqual(expectedProblems.length);
             problems.forEach((problem, index) => {
                 expect(problem.location).toEqual(
@@ -54,13 +54,10 @@ describe("lowercase_description", () => {
         });
     });
 
-    describe("method: fixLowercase description", () => {
+    describe("method: fixLowercase title", () => {
         const testData = getInvalidTestDataWithFix();
         it.each(testData)("%s", (_, text, problem, expectedFixedText) => {
-            const fixedText = LowercaseDescription.fixLowercaseDescription(
-                text,
-                problem
-            );
+            const fixedText = LowercaseTitle.fixLowercaseTitle(text, problem);
             expect(fixedText).toEqual(expectedFixedText);
         });
     });
